requirements.txt
# Core ML/CV packages
ultralytics==8.3.65
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
jinja2==3.1.4

# AWS & Monitoring
boto3==1.35.97
watchtower==3.3.0

# Data processing
pandas==2.2.3
pydantic==2.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
structlog==24.4.0
